Desk rebuild (ADR-174): the desk renders one object per work item — a state
partition strip in place of the digest window, work rows that expand into their
decision panel, run status folded into the stage chip, a single-column layout
and NAV-08, plus a body-only HITL panel extracted for reuse.

e2e: the documented baseline goes from 35 failures across 30 spec files to 1.
Closed 31 cases across 26 spec files, three of which were tests a serial-mode
skip had been hiding from every run. Four of the failures were product bugs, not
stale tests:

  * `IntegrationsPanel`, a server component, called `isManagedTokenRow` from a
    "use client" module — the project Integrations tab crashed on any project
    holding at least one API token.
  * `listProjectAutomations` sorted by a bare integer literal, which Postgres
    reads in ORDER BY as an ordinal position — `ORDER BY 0` failed to analyse
    and the Automations tab fell to the error boundary on every project.
  * `getDefaultBranch` probes with `git -C`, which resolves upward, so greenfield
    registration inside another working tree stamped that tree's default branch
    into the new project's manifest and row — for a repo gitInit then created on
    `main`.
  * The flow editor computed the canvas's refusal diagnostics and discarded
    them, showing "YAML is invalid" for a legacy manifest whose YAML parses.

Each product fix carries a regression guard that was falsified against the
unfixed code. Two documented-but-unimplemented behaviours are recorded rather
than invented: ADR-116's manifest form in the composition header, and the
unreachable `sync-notice` in the upstream-sync dialog.

Remaining: `studio-ai-assistant`, which needs a real ACP turn from the test
supervisor — infrastructure, not a selector.

+++ web/lib/scheduled-launches/queries.ts
@@ -385,7 +385,20 @@ async function listProjectAutomationRows(input: {
   const oneTimeSort: AutomationSortColumns = {
     active: sql<boolean>`${scheduledTaskLaunches.nextAttemptAt} IS NOT NULL`,
     id: sql<string>`${scheduledTaskLaunches.id}`,
-    kindRank: sql<number>`0`,
+    // `0::int`, not `0`. A BARE integer literal in ORDER BY is an ORDINAL
+    // POSITION to Postgres, not a constant — `ORDER BY 0` is out of range and
+    // the whole query dies with "ORDER BY position 0 is not in select list",
+    // which took the project Automations tab down with it. The cast makes it an
+    // expression, which is what a rank constant has to be.
+    //
+    // `1::int` below was NOT breaking anything: within one source every row has
+    // the same rank, so the key is inert, and `ORDER BY 1` happened to resolve
+    // to that query's first selected column — `runSchedules.id`, already the
+    // next tiebreaker. It is cast for the same reason regardless: as an ordinal
+    // it silently tracks the select list, so reordering the columns would
+    // change the sort. Cross-KIND ordering never came from here — it is the
+    // `kindRank(type)` comparator that merges the three sources.
+    kindRank: sql<number>`0::int`,
     nextActionAt: sql<Date | null>`${scheduledTaskLaunches.nextAttemptAt}`,
     updatedAt: sql<Date>`${scheduledTaskLaunches.updatedAt}`,
   };
@@ -396,7 +409,7 @@ async function listProjectAutomationRows(input: {
   const recurringSort: AutomationSortColumns = {
     active: sql<boolean>`${recurringNextActionAt} IS NOT NULL`,
     id: sql<string>`${runSchedules.id}`,
-    kindRank: sql<number>`1`,
+    kindRank: sql<number>`1::int`,
     nextActionAt: recurringNextActionAt,
     updatedAt: sql<Date>`${runSchedules.updatedAt}`,
   };
